web_requirements.txt
# Django Web Interface Requirements
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14.0

# WebSocket support
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0

# Redis for channels layer
redis>=5.0.0

# Static file serving
whitenoise>=6.6.0

# CORS headers for API
django-cors-headers>=4.3.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0

# HTTP client for backend API calls
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
asgiref>=3.7.0

# Utilities
python-dateutil>=2.8.0

# Development
django-debug-toolbar>=4.2.0

# Tailwind CSS integration
django-tailwind>=3.8.0
django-browser-reload>=1.12.0
